f-morseio-split # move to go module, rename cwc package to core
