+++ go/config/config.go
@@ -15,14 +15,14 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
-package cwc
+package config
 
 import (
 	"fmt"
 	"math/rand"
 
-	"../bitoip"
 	"github.com/BurntSushi/toml"
+	"github.com/G0WCZ/cwc/bitoip"
 	"github.com/golang/glog"
 )
