+++ go/core/hw/events.go
@@ -15,42 +15,21 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
-package cwc
 
-import "context"
+package hw
 
-import "../bitoip"
+import (
+	"github.com/G0WCZ/cwc/bitoip"
+	"time"
+)
 
-// This is the morse receiver
-// This polls morse input "hardware" for data
-
-func MorseRx(ctx context.Context, morseReceived chan bitoip.CarrierEventPayload, config *Config) {
-
-	ConfigureMorseInputs(config)
-
-	// Open them
-
-	// Start sampling inputs
-
+// Absolute time morse event
+type TimedBitEvent struct {
+	startTime time.Time
+	bitEvent  bitoip.BitEvent
 }
 
-// Resolve config into actual inputs that are enabled
-// For example, might be a key/keyer and text input
-func ConfigureMorseInputs(ccnfig *Config) {
-	// read input array and resolve into set of morse_in devices.
+// slice of events
+type MorseEvents []TimedBitEvent
 
-}
-
-func OpenInputs() {
-
-}
-
-func CloseInputs() {
-
-}
-
-func RunRx(ctx context.Context, morseReceived chan bitoip.CarrierEventPayload) {
-
-}
-
-// Keyer support
+//TODO conversion to/from messages
