+++ go/core/hw/register.go
@@ -16,6 +16,20 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-package io
+package hw
 
-// Keyer layer
+type InputInfo struct {
+	Name string
+}
+
+type OutputInfo struct {
+	Name string
+}
+
+type GeneralInfo struct {
+	Name string
+}
+
+var Inputs = make(map[string]InputInfo)
+var Outputs = make(map[string]OutputInfo)
+var GeneralIOs = make(map[string]GeneralInfo)
