+++ go/core/lang/morse.go
