+++ go/core/lang/morse_test.go
