+++ go/core/lang/timing.go
