+++ go/core/morse_rx.go
@@ -1,0 +1,64 @@
+/*
+Copyright (C) 2019 Graeme Sutherland, Nodestone Limited
+
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <https://www.gnu.org/licenses/>.
+*/
+package core
+
+import (
+	"context"
+	"github.com/G0WCZ/cwc/bitoip"
+	"github.com/G0WCZ/cwc/config"
+	"github.com/G0WCZ/cwc/core/hw"
+)
+
+// List of inputs to work with
+var inputs []*hw.MorseIn = make([]*hw.MorseIn, 0)
+
+// This is the morse receiver
+// This polls morse input "hardware" for data
+
+func MorseRx(ctx context.Context, morseReceived chan bitoip.CarrierEventPayload, config *config.Config) {
+
+	ConfigureMorseInputs(config)
+
+	// Open them
+
+	// Start sampling inputs
+
+}
+
+// Resolve config into actual inputs that are enabled
+// For example, might be a key/keyer and text input
+func ConfigureMorseInputs(config *config.Config) {
+	// read input array and resolve into set of morse_in devices.
+	for i, inp := range config.MorseInHardware {
+
+	}
+}
+
+func OpenInputs() {
+
+}
+
+func CloseInputs() {
+
+}
+
+func RunRx(ctx context.Context, morseReceived chan bitoip.CarrierEventPayload) {
+
+}
+
+// Keyer support
