+++ go/go.mod
@@ -1,0 +1,16 @@
+module github.com/G0WCZ/cwc
+
+go 1.12
+
+require (
+	github.com/BurntSushi/toml v0.3.1
+	github.com/creack/goselect v0.1.0 // indirect
+	github.com/gin-contrib/multitemplate v0.0.0-20190528082104-30e424939505
+	github.com/gin-gonic/gin v1.4.0
+	github.com/golang/glog v0.0.0-20160126235308-23def4e6c14b
+	github.com/google/go-cmp v0.3.0 // indirect
+	github.com/pkg/errors v0.8.1 // indirect
+	github.com/stianeikeland/go-rpio v4.2.0+incompatible
+	go.bug.st/serial.v1 v0.0.0-20180827123349-5f7892a7bb45
+	gotest.tools v2.2.0+incompatible
+)
